+++ demo/src/main/java/com/orange/ods/demo/ui/components/buttons/ButtonsContained.kt
@@ -35,7 +35,7 @@ import com.orange.ods.compose.component.list.OdsSwitchTrailing
 import com.orange.ods.compose.theme.OdsDisplaySurface
 import com.orange.ods.demo.R
 import com.orange.ods.demo.ui.components.utilities.ComponentCustomizationBottomSheetScaffold
-import com.orange.ods.demo.ui.utilities.composable.ButtonTechnicalText
+import com.orange.ods.demo.ui.utilities.composable.ButtonTechnicalTextColumn
 import com.orange.ods.demo.ui.utilities.composable.CodeImplementationColumn
 import com.orange.ods.demo.ui.utilities.composable.Subtitle
 import com.orange.ods.demo.ui.utilities.composable.Title
@@ -66,8 +66,8 @@ fun ButtonsContained(style: OdsButtonStyle) {
                     trailing = OdsSwitchTrailing(checked = leadingIcon)
                 )
                 OdsListItem(
-                    text = stringResource(id = R.string.component_button_variable_width),
-                    trailing = OdsSwitchTrailing(checked = variableWidth)
+                    text = stringResource(id = R.string.component_button_full_screen_width),
+                    trailing = OdsSwitchTrailing(checked = fullScreenWidth)
                 )
                 OdsListItem(
                     text = stringResource(id = R.string.component_state_enabled),
@@ -89,7 +89,7 @@ fun ButtonsContained(style: OdsButtonStyle) {
                     }
                 }
 
-                ContainedButton(style = containedButtonStyle.value, leadingIcon = hasLeadingIcon, enabled = isEnabled, variableWidth = hasVariableWidth)
+                ContainedButton(style = containedButtonStyle.value, leadingIcon = hasLeadingIcon, enabled = isEnabled, fullScreenWidth = hasFullScreenWidth)
 
                 Spacer(modifier = Modifier.padding(top = dimensionResource(R.dimen.spacing_s)))
 
@@ -98,17 +98,17 @@ fun ButtonsContained(style: OdsButtonStyle) {
                         style = containedButtonStyle.value,
                         leadingIcon = hasLeadingIcon,
                         enabled = isEnabled,
-                        variableWidth = hasVariableWidth,
+                        fullScreenWidth = hasFullScreenWidth,
                         displaySurface = displaySurface
                     )
                 }
 
                 CodeImplementationColumn {
-                    ButtonTechnicalText(
+                    ButtonTechnicalTextColumn(
                         componentName = OdsComponent.OdsButton.name,
                         style = containedButtonStyle.value.fullName,
                         enabled = isEnabled,
-                        variableWidth = hasVariableWidth,
+                        fullScreenWidth = hasFullScreenWidth,
                         icon = hasLeadingIcon
                     )
                 }
@@ -122,13 +122,13 @@ private fun ContainedButton(
     style: OdsButtonStyle,
     leadingIcon: Boolean,
     enabled: Boolean,
-    variableWidth: Boolean,
+    fullScreenWidth: Boolean,
     displaySurface: OdsDisplaySurface = OdsDisplaySurface.Default
 ) {
     val modifier = Modifier.padding(horizontal = dimensionResource(R.dimen.screen_horizontal_margin), vertical = dimensionResource(R.dimen.spacing_m))
 
     OdsButton(
-        modifier = if (variableWidth) modifier else modifier.fillMaxWidth(),
+        modifier = if (fullScreenWidth) modifier.fillMaxWidth() else modifier,
         icon = if (leadingIcon) painterResource(id = R.drawable.ic_coffee) else null,
         text = stringResource(if (enabled) R.string.component_state_enabled else R.string.component_state_disabled),
         onClick = {},